Split iOS flashcards domain types
