+++ apps/android/data/local/src/main/java/com/flashcardsopensourceapp/data/local/model/FlashcardsModels.kt
@@ -3,7 +3,14 @@ package com.flashcardsopensourceapp.data.local.model
 /*
  Keep these shared data contracts aligned with:
  - apps/web/src/types.ts
- - apps/ios/Flashcards/Flashcards/FlashcardsTypes.swift
+ - apps/ios/Flashcards/Flashcards/Cards/CardDeckTypes.swift
+ - apps/ios/Flashcards/Flashcards/Cards/CardDeckSnapshotTypes.swift
+ - apps/ios/Flashcards/Flashcards/Cards/TagTypes.swift
+ - apps/ios/Flashcards/Flashcards/Cloud/CloudTypes.swift
+ - apps/ios/Flashcards/Flashcards/Cloud/Sync/CloudSyncTypes.swift
+ - apps/ios/Flashcards/Flashcards/Database/WorkspaceLocalTypes.swift
+ - apps/ios/Flashcards/Flashcards/Review/FsrsTypes.swift
+ - apps/ios/Flashcards/Flashcards/Review/ReviewTypes.swift
  */
 
 data class WorkspaceSummary(
@@ -363,7 +370,7 @@ data class PersistedOutboxEntry(
     val operation: SyncOperation
 )
 
-// Keep in sync with apps/backend/src/schedule.ts::FsrsCardState, apps/web/src/types.ts::FsrsCardState, and apps/ios/Flashcards/Flashcards/FlashcardsTypes.swift::FsrsCardState.
+// Keep in sync with apps/backend/src/schedule.ts::FsrsCardState, apps/web/src/types.ts::FsrsCardState, and apps/ios/Flashcards/Flashcards/Review/FsrsTypes.swift::FsrsCardState.
 enum class FsrsCardState {
     NEW,
     LEARNING,
@@ -401,7 +408,7 @@ data class DeckSummary(
     val updatedAtMillis: Long
 )
 
-// Keep in sync with apps/backend/src/workspaceSchedulerSettings.ts::WorkspaceSchedulerSettings, apps/web/src/types.ts::WorkspaceSchedulerSettings, and apps/ios/Flashcards/Flashcards/FlashcardsTypes.swift::WorkspaceSchedulerSettings.
+// Keep in sync with apps/backend/src/workspaceSchedulerSettings.ts::WorkspaceSchedulerSettings, apps/web/src/types.ts::WorkspaceSchedulerSettings, and apps/ios/Flashcards/Flashcards/Review/FsrsTypes.swift::WorkspaceSchedulerSettings.
 data class WorkspaceSchedulerSettings(
     val workspaceId: String,
     val algorithm: String,
@@ -413,7 +420,7 @@ data class WorkspaceSchedulerSettings(
     val updatedAtMillis: Long
 )
 
-// Keep in sync with apps/backend/src/cards.ts::Card, apps/web/src/types.ts::Card, and apps/ios/Flashcards/Flashcards/FlashcardsTypes.swift::Card.
+// Keep in sync with apps/backend/src/cards.ts::Card, apps/web/src/types.ts::Card, and apps/ios/Flashcards/Flashcards/Cards/CardDeckTypes.swift::Card.
 data class CardSummary(
     val cardId: String,
     val workspaceId: String,
