+++ apps/ios/Flashcards/Flashcards/Cards/CardPresentationTypes.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+enum CardsPresentationRequest: Hashable, Sendable {
+    case createCard
+}
