+++ apps/web/src/types.ts
@@ -10,12 +10,13 @@ import type { Locale } from "./i18n/types";
  * Keep these FSRS-facing types aligned with:
  * - apps/backend/src/schedule.ts
  * - apps/backend/src/workspaceSchedulerSettings.ts
- * - apps/ios/Flashcards/Flashcards/FlashcardsTypes.swift
+ * - apps/ios/Flashcards/Flashcards/Cards/CardDeckTypes.swift
+ * - apps/ios/Flashcards/Flashcards/Review/FsrsTypes.swift
  * - apps/android/data/local/src/main/java/com/flashcardsopensourceapp/data/local/model/FlashcardsModels.kt
  * - docs/fsrs-scheduling-logic.md
  */
 export type EffortLevel = "fast" | "medium" | "long";
-// Keep in sync with apps/backend/src/schedule.ts::FsrsCardState, apps/ios/Flashcards/Flashcards/FlashcardsTypes.swift::FsrsCardState, and apps/android/data/local/src/main/java/com/flashcardsopensourceapp/data/local/model/FlashcardsModels.kt::FsrsCardState.
+// Keep in sync with apps/backend/src/schedule.ts::FsrsCardState, apps/ios/Flashcards/Flashcards/Review/FsrsTypes.swift::FsrsCardState, and apps/android/data/local/src/main/java/com/flashcardsopensourceapp/data/local/model/FlashcardsModels.kt::FsrsCardState.
 export type FsrsCardState = "new" | "learning" | "review" | "relearning";
 
 export type CardFilter = Readonly<{
@@ -395,7 +396,7 @@ export type HomeSnapshot = Readonly<{
   reviewedCount: number;
 }>;
 
-// Keep in sync with apps/backend/src/cards.ts::Card, apps/ios/Flashcards/Flashcards/FlashcardsTypes.swift::Card, and apps/android/data/local/src/main/java/com/flashcardsopensourceapp/data/local/model/FlashcardsModels.kt::CardSummary.
+// Keep in sync with apps/backend/src/cards.ts::Card, apps/ios/Flashcards/Flashcards/Cards/CardDeckTypes.swift::Card, and apps/android/data/local/src/main/java/com/flashcardsopensourceapp/data/local/model/FlashcardsModels.kt::CardSummary.
 export type Card = Readonly<{
   cardId: string;
   frontText: string;
@@ -521,7 +522,7 @@ export type TagSuggestion =
     cardsCount: number;
   }>;
 
-// Keep in sync with apps/ios/Flashcards/Flashcards/FlashcardsTypes.swift::WorkspaceSchedulerSettings and apps/backend/src/workspaceSchedulerSettings.ts::WorkspaceSchedulerSettings.
+// Keep in sync with apps/ios/Flashcards/Flashcards/Review/FsrsTypes.swift::WorkspaceSchedulerSettings and apps/backend/src/workspaceSchedulerSettings.ts::WorkspaceSchedulerSettings.
 export type WorkspaceSchedulerSettings = Readonly<{
   algorithm: "fsrs-6";
   desiredRetention: number;
